Move map logic to TiledMapLoader.java
