+++ core/src/model/ModelFactory.java
@@ -3,7 +3,7 @@ package model;
 public class ModelFactory {
     public static Model makeModel(String mapName) {
         Model model = new Model();
-        model.initialize(mapName);
+        model.initialize(new TiledMapLoader(mapName));
         return model;
     }
 }
